refactor(job_attachments)!: Move class definitions from `_utils` to `models.py`
BREAKING CHANGE:
- Relocated the classes (`AssetLoadingMethod`, `FileConflictResolution`, `FileSystemLocationType`, `OperatingSystemFamily`) from _utils to models.py
